added javascript to make the optional characters left work
this took a lot of trial and error, and it also required me to hardcode minimal height of the cheepbox into the cshtml instead of the css document. as css document didnt seem to change anything. also i used chatgpt to try and figure out a way to do it without javascript but it would make the function non dynamic and kinda useless, so i gave up on that, but i was still inspired by its ideas so i put it as a co-auth