Changed Author table to consist of Authors instead of IdentityUsers.
This broke most of our tests along with our database, and required a lot of messing around with SQLite DB sources. Especially a bunch of missing colons on some connections strings. During this we also realized a big bug about simultaneous access to the database, which arose from our tests. We got the fix from a StackOverflow post.

Co-authored-by: Adam <adni@itu.dk>
Co-authored-by: Shain Dohan (StackOverflow)

https://stackoverflow.com/questions/51080604/no-such-table-ef-core-with-sqlite-in-memory