Removed change username and add password for github login
You are no longer able to change your username when logging in with github or set a password. If your username or email exists in the database you will receive an error message. The author now have two additional fields for authentication with github since they are being used for login, so you would be able to change the username if it where to be added. The field with a list of cheeps has been added to the author domain.